// frontend/src/app/features/dashboard/models/dashboard-widget-content.model.ts
import { AnalysisPriorityValue } from './analysis-priority';
import { AnalysisStatusValue } from './analysis-status';
import { GroupByValue } from './group-by';

export interface DashboardWidgetFilters {
  dateFrom?: string;
  dateTo?: string;
  laboratoryId?: number;
  analysisCategoryId?: number;
  priority?: AnalysisPriorityValue;
  status?: AnalysisStatusValue;
}

export interface DashboardWidgetContent {
  filters?: DashboardWidgetFilters;
  groupBy?: GroupByValue;
}